+++ src/types/schemas/index.ts
@@ -1,2 +1,2 @@
 export * from "./a2a/index.js";
-// export * from "./deployment/index.js";
+export * from "./deployment/index.js";